+++ weblate/trans/views/basic.py
@@ -42,6 +42,7 @@ from weblate.trans.forms import (
     AutoForm, ReviewForm, get_new_language_form,
     ReportsForm, ReplaceForm, NewUnitForm, BulkStateForm, DownloadForm,
     DeleteForm, ProjectRenameForm, ComponentRenameForm, ComponentMoveForm,
+    WhiteboardForm,
 )
 from weblate.accounts.notifications import notify_new_language
 from weblate.utils.views import (
@@ -141,6 +142,9 @@ def show_project(request, project):
                 translation__component__project=obj
             ).distinct().count(),
             'search_form': SearchForm(),
+            'whiteboard_form': optional_form(
+                WhiteboardForm, user, 'project.edit', obj
+            ),
             'delete_form': optional_form(
                 DeleteForm, user, 'project.edit', obj, obj=obj
             ),
@@ -191,6 +195,9 @@ def show_component(request, project, component):
             'mass_state_form': optional_form(
                 BulkStateForm, user, 'translation.auto', obj,
                 user=user, obj=obj
+            ),
+            'whiteboard_form': optional_form(
+                WhiteboardForm, user, 'component.edit', obj
             ),
             'delete_form': optional_form(
                 DeleteForm, user, 'component.edit', obj, obj=obj
@@ -254,6 +261,9 @@ def show_translation(request, project, component, lang):
                 initial={
                     'value': Unit(translation=obj, id_hash=-1),
                 },
+            ),
+            'whiteboard_form': optional_form(
+                WhiteboardForm, user, 'component.edit', obj
             ),
             'delete_form': optional_form(
                 DeleteForm, user, 'translation.delete', obj, obj=obj